C#: Reduce disabled nullability regions by splitting 'Extractor' and 'Analyser'
