+++ csharp/extractor/Semmle.Extraction.CIL/Analyser.cs
@@ -7,7 +7,7 @@ namespace Semmle.Extraction.CIL
 {
     public static class Analyser
     {
-        private static void ExtractCIL(Extractor extractor, TrapWriter trapWriter, bool extractPdbs)
+        private static void ExtractCIL(NonStandaloneExtractor extractor, TrapWriter trapWriter, bool extractPdbs)
         {
             using var cilContext = new Context(extractor, trapWriter, extractor.OutputPath, extractPdbs);
             cilContext.Populate(new Assembly(cilContext));
@@ -33,7 +33,7 @@ namespace Semmle.Extraction.CIL
             {
                 var canonicalPathCache = CanonicalPathCache.Create(logger, 1000);
                 var pathTransformer = new PathTransformer(canonicalPathCache);
-                var extractor = new Extractor(assemblyPath, logger, pathTransformer);
+                var extractor = new NonStandaloneExtractor(assemblyPath, logger, pathTransformer);
                 var transformedAssemblyPath = pathTransformer.Transform(assemblyPath);
                 var project = layout.LookupProjectOrDefault(transformedAssemblyPath);
                 using var trapWriter = project.CreateTrapWriter(logger, transformedAssemblyPath.WithSuffix(".cil"), trapCompression, discardDuplicates: true);
