+++ csharp/extractor/Semmle.Extraction.CSharp/Extractor/AnalysisAction.cs
@@ -1,0 +1,12 @@
+namespace Semmle.Extraction.CSharp
+{
+    /// <summary>
+    /// What action was performed when extracting a file.
+    /// </summary>
+    public enum AnalysisAction
+    {
+        Extracted,
+        UpToDate,
+        Excluded
+    }
+}
