+++ csharp/extractor/Semmle.Extraction/Extractor/Extractor.cs
@@ -1,5 +1,4 @@
 using System.Collections.Generic;
-using Microsoft.CodeAnalysis;
 using Semmle.Util.Logging;
 
 namespace Semmle.Extraction
@@ -7,21 +6,17 @@ namespace Semmle.Extraction
     /// <summary>
     /// Implementation of the main extractor state.
     /// </summary>
-    public class Extractor
+    public abstract class Extractor
     {
-        public bool Standalone => OutputPath is null;
-
-        public string OutputPath { get; }
+        public abstract bool Standalone { get; }
 
         /// <summary>
         /// Creates a new extractor instance for one compilation unit.
         /// </summary>
-        /// <param name="outputPath">The name of the output DLL/EXE, or null if not specified (standalone extraction).</param>
         /// <param name="logger">The object used for logging.</param>
         /// <param name="pathTransformer">The object used for path transformations.</param>
-        public Extractor(string outputPath, ILogger logger, PathTransformer pathTransformer)
+        protected Extractor(ILogger logger, PathTransformer pathTransformer)
         {
-            OutputPath = outputPath;
             Logger = logger;
             PathTransformer = pathTransformer;
         }
